# More text layout optimizations