… src/Avalonia.Base/Media/TextFormatting/TextLayout.cs
… src/Avalonia.Base/Media/TextFormatting/TextLineImpl.cs
… src/Avalonia.Base/Media/TextFormatting/Unicode/BiDiAlgorithm.cs
… src/Avalonia.Base/Media/TextFormatting/Unicode/BiDiData.cs
… src/Avalonia.Base/Media/TextFormatting/Unicode/Codepoint.cs
… src/Avalonia.Base/Media/TextFormatting/Unicode/CodepointEnumerator.cs
… src/Avalonia.Base/Media/TextFormatting/Unicode/GraphemeEnumerator.cs
… src/Avalonia.Base/Media/TextFormatting/Unicode/LineBreakEnumerator.cs
… src/Avalonia.Controls/TextBox.cs
… src/Skia/Avalonia.Skia/Avalonia.Skia.csproj
… src/Skia/Avalonia.Skia/TextShaperImpl.cs
… src/Windows/Avalonia.Direct2D1/Media/TextShaperImpl.cs
… tests/Avalonia.Base.UnitTests/Media/TextFormatting/GraphemeBreakClassTrieGeneratorTests.cs
… tests/Avalonia.Base.UnitTests/Media/TextFormatting/LineBreakEnumuratorTests.cs
… tests/Avalonia.Benchmarks/Avalonia.Benchmarks.csproj
… tests/Avalonia.Benchmarks/Text/HugeTextLayout.cs
… tests/Avalonia.Skia.UnitTests/Media/TextFormatting/TextFormatterTests.cs
… tests/Avalonia.Skia.UnitTests/Media/TextFormatting/TextLayoutTests.cs
… tests/Avalonia.UnitTests/HarfBuzzTextShaperImpl.cs
… tests/Avalonia.UnitTests/MockGlyphRun.cs
… tests/Avalonia.UnitTests/MockTextShaperImpl.cs