# What does the pull request do?

Episode 3: Revenge of the Codepoint!

Here's another round of text layout performance improvements, this time focused on optimizing the core of text processing: `Codepoint`, `Grapheme` and their enumerators. These are called so often in the text layout process that even micro optimizations are significant here.

- `Codepoint.ReadAt` is optimized to have less branches, less bound checks added by the JIT, `ReplacementCodepoint` being inlined as const instead of loading a static field, and the whole method being aggressively inlined. It's the very core of text processing, and these simple changes have measurable impact.

- `GraphemeEnumerator` is improved to avoid reading codepoints twice. Previously n+1 codepoints were read to form a grapheme with n codepoints. The extra codepoint acting as a boundary was discarded, then the next grapheme was read, starting again at that codepoint. This PR changes the algorithm so that the extra codepoint is used in the next grapheme instead of having to be read again. This cuts the time spent reading codepoints in half for strings without surrogates.

- `LineBreakEnumerator`, `BidiAlgorithm` and `BidiData` use bitmasks instead of many if/switch when possible to avoid many comparisons.

- At a bit higher level, `FontManager` and `GlyphTypeface` are stored and passed around to avoid having to lookup `FontManager.Current` in the container many times.

- The `Grapheme` changes from #9928 (removal of the `Text` field) are restored, they were lost with the changes in #9982.

- Some `IReadOnlyList<T>` interfaces calls are avoided.

# Perf changes

|                Method |  Wrap |  Trim |      Before |    After   | Change |
|---------------------- |------ |------ |------------:|-----------:|-------:|
|       BuildTextLayout | False | False |    340.5 us |   206.2 us |   -39% |
| BuildEmojisTextLayout | False | False |    735.7 us |   483.3 us |   -34% |
|   BuildManySmallTexts | False | False |  2,579.7 us | 1,958.8 us |   -24% |
|  VirtualizeTextBlocks | False | False |  3,784.5 us | 3,174.9 us |   -16% |
|       BuildTextLayout | False |  True |    362.3 us |   219.5 us |   -39% |
| BuildEmojisTextLayout | False |  True |    773.1 us |   517.8 us |   -33% |
|   BuildManySmallTexts | False |  True |  2,928.5 us | 2,135.9 us |   -27% |
|  VirtualizeTextBlocks | False |  True |  3,772.2 us | 3,207.1 us |   -15% |
|       BuildTextLayout |  True | False |  2,249.2 us | 1,541.5 us |   -31% |
| BuildEmojisTextLayout |  True | False | 11,478.4 us | 8,212.8 us |   -28% |
|   BuildManySmallTexts |  True | False |  3,490.3 us | 2,637.5 us |   -24% |
|  VirtualizeTextBlocks |  True | False |  3,705.2 us | 3,125.1 us |   -16% |
|       BuildTextLayout |  True |  True |  2,284.8 us | 1,558.6 us |   -32% |
| BuildEmojisTextLayout |  True |  True | 11,148.2 us | 8,287.4 us |   -26% |
|   BuildManySmallTexts |  True |  True |  3,711.1 us | 2,973.4 us |   -20% |
|  VirtualizeTextBlocks |  True |  True |  3,798.6 us | 3,117.0 us |   -18% |

The benchmarks were executed on master and this branch with `MockGlyphRun` a bit optimized to avoid LINQ. Mocks aren't meant to be optimized, but it was taking a significant time.

While executing with mocks is nice to see how much the code was improved, I also added an option to execute the benchmark with the Skia renderer. Results are a bit less shiny, but more accurately represent a real usage. HarfBuzz takes most of the time here, which is great (Avalonia shouldn't be the bottleneck).

<details>
  <summary>Results with HarfBuzz</summary>

|                Method |  Wrap |  Trim |      Before |      After | Change |
|---------------------- |------ |------ |------------:|-----------:|-------:|
|       BuildTextLayout | False | False |    758.7 us |   619.3 us |   -18% |
| BuildEmojisTextLayout | False | False |  2,079.1 us | 1,858.9 us |   -11% |
|   BuildManySmallTexts | False | False |  5,564.0 us | 5,085.6 us |    -9% |
|  VirtualizeTextBlocks | False | False |  7,104.4 us | 6,445.8 us |    -9% |
|       BuildTextLayout | False |  True |    813.5 us |   667.5 us |   -18% |
| BuildEmojisTextLayout | False |  True |  2,195.6 us | 1,926.7 us |   -12% |
|   BuildManySmallTexts | False |  True |  5,636.6 us | 4,999.8 us |   -11% |
|  VirtualizeTextBlocks | False |  True |  7,235.0 us | 6,432.8 us |   -11% |
|       BuildTextLayout |  True | False |  1,811.1 us | 1,278.0 us |   -29% |
| BuildEmojisTextLayout |  True | False |  5,412.9 us | 3,888.3 us |   -28% |
|   BuildManySmallTexts |  True | False |  6,645.3 us | 5,772.4 us |   -13% |
|  VirtualizeTextBlocks |  True | False |  7,121.7 us | 6,290.1 us |   -12% |
|       BuildTextLayout |  True |  True |  1,797.3 us | 1,277.2 us |   -29% |
| BuildEmojisTextLayout |  True |  True |  5,609.6 us | 3,872.6 us |   -31% |
|   BuildManySmallTexts |  True |  True |  6,530.3 us | 5,778.6 us |   -12% |
|  VirtualizeTextBlocks |  True |  True |  7,096.7 us | 6,498.6 us |    -8% |

</details>